created scale.js, refactoring
